Updated README for documentation of location field
The current documentation says that the location will have the fully qualified domain name. The FQDN terminology is appropriate for server names and not for file locations/paths.
File paths are expressed as URLs. Proposed a change in the documentation to use correct terminology and provide examples to clarify that the file paths can be local, relative or remote.